Wrote an initial high concept blurb in the Bluesky folder.
I'm grounding the story for my ride on the chasse gallerie for the previously mentioned opportunities to align a diegetic vehicle with a ride system. Using the wrapper premise of a deal with the devil I think also sets up the idea that you're being chased or impeded not just by the Devil himself, but also by any number of monsters I'd like to include from other Quebcois folklore. Even without fully sketching out storybeats yet, I can see a vague path of a ride that starts as a flume, takes off into a suspended coaster, zips past a bunch of monsters, and then maybe ends with a splashdown and return to the loading station.