Designed a background for final submission document.
This is getting a little ahead of myself, but because I know I want the final submission of designs to appear on 11X17 sheets, I might as well start working in those dimensions and I had a pretty good sense that I wanted the BG to look sorta like singed parchement to match both the period of the ride and the fact that the devil features prominently in it. I used a stock photo of parchment-like paper I made years ago and just overlapped it and flipped it around a lot with feathered edges to mask that it was the same photo over and over. Added a singed edge texture and I'm pretty happy with the outcome. Might have to lighten it a touch for legibility with text/lineart.