Updated to-to list.
Given I have less than a week left, it's going to be a real mad-dash to produce deliverables. To that end, I'm sorting out precisely what documents I want to be able to deliver. Rather than illustrate every storybeat of the ride, I've opted to choose the five most evocative and illustrate those so the emotional arc is conveyed.