Brought some photos of voyageur canoes into a document to assess salient visual features.
Based on analysis, three main visual features to capture in the ride vehicle, a birch bark texure for missle cars of the ride train, and rounded nose and tail with a white patch and design for the lead and tail cars. I'm a little surprised just how consisting depictions and recreations of these canoes are. I wonder if this is a case of a handful of famous examples set the model and then every recreation simply adds weight to those one or two examples.