Using a tracing of one ride vehicle from a photo of the Vortex ride at Canada's wonderland, created a basic line art of the six car vehicle in perspective.
Next step will be to make adjustments to achieve the canoe profile I want. Looking at the finished product, I actually think my cars in perspective are slightly misaligned, but I don't have a lot of time to work on this and I do think it's close enough to be reasonably convincing.