Designed most of the paint scheme for the ride vehicle.
Had to stop for a few hours in the middle because my computer started having repeated kernel panics. Not really sure why, but this does make me reflect on the fragility of MDM to things like a lack of wifi or a computer crash. As I sat, computerless, for a few hours, I found myself considering how I might, if at all, complete my comp project and the answer wasn't super clear. Like even if I'd wanted to continue with analog tools (i.e. pencil and paper), how would I have documented it for the purposes of MDM?