Finished up the paint scheme for the ride vehicle.
Decided the white patches with decorative motifs for the head and tail of the train should have a devil's head and tail as a tie-in to the narrative premise of the devil making the canoe fly. I could go with a more traditional floral or geometric motif, but I kinda like the idea of the devil being literally in the ride vehicle's details.