Added a visual analysis of reference photos.
Again, found fairly convergent features that should make deriving a operator uniform pretty straightforward. Though generally pretty simple, there does some to be some level of layering, so part of the costume design will definitely be how to reduce the number of individual articles down to a minimum for simplicity. For example, pants or leg wraps, seem to be gartered just below the knee, but I'm thinking the costume design should probably incorporate the garters as a stretch band sewn to the pants. The fact that both warm coats and light shirts were common parts of fur trader garb does make it easy though to design costumes with an optional overlayer for inclement weather.