Started a file for designing a logo for the ride.
I think for this logo I want to aim at some that feels old-script-y and a little bit foreboding, without overly telegraphing the hellish tones of the ride. Maybe some flame motifs, but otherwise no visible devils as I think this tends to be the go-to for rides named after the devil and is a bit cheesy in my opinion. I do want to call the ride The Devil's Flight though to evoke some of that flavor and also because the french name of the legend Chasse Galerie means nothing to an anglophone audience or even a french audience that doesn't know the legend already (the name doesn't really describe the story at all.)