Added some text, including callouts to details in the garment.
I'm also think I might want to depict the satchel in a separate corner because the way I've posed the model, the satchel is likely to either be hidden behind the right arm, or be out of view on the left side of the body. Luckily I do still have an open corner to the page.