Finished colouring the uniform.
I've decided in the end to not feature the satchel at all. With more time and preparing more documents, I would, but at this stage I think I actually want to save the space for fabric swatches. If I were going to include the satchel, I'd probably want to provide a detailed pattern for stitching/beadwork on it, and likewise might provide a detail for the woven sash's pattern.