Sketched out my concept art and started photobash/painting in the details.
Probably should have pushed a reflection after the sketch but I'm not sure there's much to say about it and I've saved the sketch as a layer of the file. The main challenge with the photobashing is going to be collaging pieces together, then making them all look cohesive as if they're from the same environment/photo. For example, I'm already having to paint snow onto some of the trees that weren't snowy enough as they were.