Finished the concept art for the queue page.
Took me a little while because I took a break in the middle for dinner. Maybe I should have pushed a comit in that time but eh, here we are. Sitting down to do photobashing and digital painting always gives me pause until I do it and remember I've actually sorta gotten good at it over the years. Feeling less daunted about all the concept art I tasked myself with for the storybeats.