Sketched out some general thoughts about track layout and initialized a file for desiging the actual layout.
Mostly just translating my textual description of the ride into concrete contours. It's still a bit sketchy, which will be for me to work out in the next stage, but I tried graphing the general height over time to make it a bit clearer in my mind what the height should be at any given point in time. Also got a rough sense of what sections have to be straight and which bank a bit, and which are fully windy so I can start piecing them together into a layout.