Made an elevation view of the track.
Had to redo a few sections because I kept forgetting that a relatively flat part still isn't going to be perfectly flat and so a winding path will look like a zig zag in profile. Once I had it initially sketch, I also realized it looked a bit boring so I added a few dips and climbs to the more tense parts. I sorta figure that this is, after all, a dark ride and not a full coaster, and because the ride cars for this coaster system swing, they don't need a lot of intense twists and turns to get some fair motion going.