Finished the concept art for the second storybeats page.
Decided to cut the werewolf entirely for the art but added a caption in the margin to really emphasize the opposing motion of the ride car and the falling trees since those are the "clashing effects" I'm trying to convey more so than the monster causing it. Also pretty happy with how the paintover went to make the real train look like my vehicle design. Also painted over some of the faces to make them look a bit more horrified by what's happening.