Started making the concept art for the first storybeats page.
I want to go with depicting the moment the ride vehicle goes over the waterfall as I feel this juxtaposes senses of falling and flying while also creating a triumphant resolution to fear, both tying into the brief theme of clashing affect and effect. So far I've just begun collaging pieces of actual photos of the ride system together to block in the space.