Finished the train paint over.
Just want to do a final pass on the faces now to change the expressions a bit and anonymize faces. A bit on the fence as to whether or not the lighting in the scene is acceptablee—values are a bit muddy—but I think I might be satisfied calling it done for the time being.