Finally finished up the first storybeats page.
As I look towards the final illustration, I'm wondering if the splashdown would be too similar to the waterfall scene. Maybe it would be better to depict the dragged to hell scene instead? If so, I'm not entirely sure if it would be best to show from like a zoomed out cut-away of what would effectively be an enclosed lift hill, or if it would be more compelling from a rider perspective with the light at the end of the tunnel.