Dropped in some photos of the ride system for paintover and did some quick blockout for the sake of composiiton.
I decided to go with the hell scene after all and a rider POV for a bit of interest. Given the perspective, I also tracked down a reference from one car back to ensure the ride system is still somewhat visible n the composition. Not entirely sure how this is going to turn out, but I'm gonna trust the process at this stage.