Established lighting and color in the scene.
Might want to go back in with some yellower highlights to give the vehicle some texture, but given the lighting in such a tunnel would be pretty monochrome, I'm not entirely worried. Now I just need to design the walls in the tunnel and paint the dashboard of the the car the viewer is in.