Finished up the art for storybeats 3.
Not much to say. Went with something simple for the walls—fire and grasping hands. This is partially a time consideration but also I feel like it's less likely to feel goofy than an overdesigned host of demons. That's the last of my deliverables, so now I just have to prep everything for my presentation tomorrow.